% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_time.R
\name{thermal_time}
\alias{thermal_time}
\title{Calculate thermal time for a crop}
\usage{
thermal_time(..., crop)
}
\arguments{
\item{...}{Arguments passed to the selected crop's thermal time function.}

\item{crop}{The crop model to use. Supported crops are \code{wheat}, \code{canola},
\code{chickpea}, \code{lupin}, \code{fababean}, and \code{lentil}.}
}
\value{
The thermal time.
}
\description{
Calculate thermal time for a crop
}
\examples{
mint <- c(0, 10)
maxt <- c(30, 40)
date <- as.Date(c("2024-01-01", "2024-01-02"))
thermal_time(mint, maxt, date = date, latitude = -27, crop = "lupin")
}
