% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermal_time.R
\name{thermal_time}
\alias{thermal_time}
\title{Calculate thermal time for a crop}
\usage{
thermal_time(..., crop)
}
\arguments{
\item{...}{Arguments passed to the selected crop's thermal time function.}

\item{crop}{The crop model to use. Supported crops are \code{wheat}, \code{canola},
\code{chickpea}, \code{lupin}, \code{fababean}, and \code{lentil}.}
}
\value{
The thermal time.
}
\description{
Calculate thermal time for a crop
}
\examples{

library(apsimwise)
# Example weather data
mint <- c(0, 10)
maxt <- c(30, 40)
date <- as.Date(c("2024-01-01", "2024-01-02"))

# Use default lupin parameters
thermal_time(
  mint,
  maxt,
  date = date,
  lat = -27,
  crop = "lupin"
)

# Modify the global lupin thermal time response for all following calculations
crop("lupin")$set(
  "phenology.thermal_time" =
  list(
    x = c(0, 25, 35),
    y = c(0, 25, 0)
  )
)
thermal_time(
  mint,
  maxt,
  date = date,
  lat = -27,
  crop = "lupin"
)

# Then the parameters can be reset
crop("lupin")$reset()


# Override thermal time parameters for a single call
thermal_time(
  mint,
  maxt,
  date = date,
  lat = -27,
  crop = "lupin",
  x_temp = c(0, 25, 35),
  y_temp = c(0, 25, 0)
)

# Calculate thermal time from a weather file
met_file <- system.file("extdata/ppd_72150.met", package = "tidyweather")
# Read weather records
records <- read_weather(met_file)
# Calculate thermal time for lupin using tidyverse ways
library(dplyr)
records |>
    mutate(
        tt = thermal_time(
            mint = mint,
            maxt = maxt,
            date = date,
            latitude = latitude[1],
            crop = "lupin"
        )
    ) |>
    select(name, date, mint, maxt, tt)
# Calculate thermal time for lupin grouped by year and site if there are multiple sites
# Create some fake data for a second site
records2 <- records |>
    mutate(
        name = "site2",
        mint = mint + 1,
        maxt = maxt + 1,
        latitude = latitude - 10
    ) |>
    bind_rows(records)
# Calculate thermal time for lupin grouped by site
records2 |>
    group_by(name) |>
    mutate(
        tt = thermal_time(
            mint = mint,
            maxt = maxt,
            date = date,
            latitude = latitude[1],
            crop = "lupin"
        )
    ) |>
    select(name, date, mint, maxt, tt)
}
