% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vernalisation_time.R
\name{vernalisation_days}
\alias{vernalisation_days}
\title{Calculate vernalisation days for a crop}
\usage{
vernalisation_days(..., crop)
}
\arguments{
\item{...}{Arguments passed to the selected crop's vernalisation days function.}

\item{crop}{The crop model to use. Supported crops are \code{canola},
\code{chickpea}, \code{lupin}, \code{fababean}, and \code{lentil}.}
}
\value{
The vernalisation days.
}
\description{
Calculate vernalisation days for a crop
}
\examples{
mint <- c(0, 10)
maxt <- c(30, 40)
vernalisation_days(mint, maxt, crop = "lupin")
}
